requirements.txt
Click==7.0
Flask==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
openslide-python==1.1.1
Pillow==7.0.0
Werkzeug==1.0.0